+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - API `/api/convert/pandoc` 兼容思源版本 `v2.9.3` | API `/api/convert/pandoc` is compatible with SiYuan version `v2.9.3`
 - 添加 API `/api/file/getFile` | Add API `/api/file/getFile`
 - 添加 API `/api/file/putFile` | Add API `/api/file/putFile`
+- 优化 API `/api/convert/pandoc` 测试 | Optimize API `/api/convert/pandoc` test
 
 ## v0.0.2 / 2023-06-26